services/error-service/pyproject.toml
[project]
name = "error-service"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.3",
    "flask-cors>=6.0.2",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=25.1.0",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
]
